Move gazebo plugins to the scenario folder
