+++ cpp/scenario/plugins/ECMProvider/ECMSingleton.cpp
@@ -1,19 +1,37 @@
 /*
- * Copyright (C) 2019 Istituto Italiano di Tecnologia (IIT)
+ * Copyright (C) 2020 Istituto Italiano di Tecnologia (IIT)
  * All rights reserved.
+ *
+ * This project is dual licensed under LGPL v2.1+ or Apache License.
+ *
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
  *
  * This software may be modified and distributed under the terms of the
  * GNU Lesser General Public License v2.1 or any later version.
+ *
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
-#include "gympp/gazebo/ECMSingleton.h"
+#include "scenario/plugins/gazebo/ECMSingleton.h"
 #include "gympp/base/Log.h"
 
 #include <mutex>
 #include <ostream>
 #include <unordered_map>
 
-using namespace gympp::gazebo;
+using namespace scenario::plugins::gazebo;
 using RobotName = std::string;
 
 struct Pointers
@@ -41,7 +59,7 @@ public:
 };
 
 ECMSingleton::ECMSingleton()
-    : pImpl{new Impl(), [](Impl* impl) { delete impl; }}
+    : pImpl{new Impl()}
 {}
 
 ECMSingleton& ECMSingleton::get()
