+++ cpp/scenario/plugins/Physics/Physics.h
@@ -1,13 +1,6 @@
 /*
- * Copyright (C) 2019 Istituto Italiano di Tecnologia (IIT)
+ * Copyright (C) 2020 Open Source Robotics Foundation
  * All rights reserved.
- *
- * This software may be modified and distributed under the terms of the
- * GNU Lesser General Public License v2.1 or any later version.
- *
- * ==================================================
- *
- * Copyright (C) 2018 Open Source Robotics Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,22 +13,23 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
- *
  */
 
-#ifndef GYMPP_PLUGINS_PHYSICS
-#define GYMPP_PLUGINS_PHYSICS
+#ifndef SCENARIO_PLUGINS_GAZEBO_PHYSICS
+#define SCENARIO_PLUGINS_GAZEBO_PHYSICS
 
 #include <ignition/gazebo/System.hh>
 #include <memory>
 
-namespace gympp {
+namespace scenario {
     namespace plugins {
-        class Physics;
+        namespace gazebo {
+            class Physics;
+        } // namespace gazebo
     } // namespace plugins
-} // namespace gympp
+} // namespace scenario
 
-class gympp::plugins::Physics final
+class scenario::plugins::gazebo::Physics final
     : public ignition::gazebo::System
     , public ignition::gazebo::ISystemUpdate
 {
@@ -51,4 +45,4 @@ private:
     std::unique_ptr<Impl> pImpl;
 };
 
-#endif // GYMPP_PLUGINS_PHYSICS
+#endif // SCENARIO_PLUGINS_GAZEBO_PHYSICS
