+++ cpp/scenario/plugins/RobotController/RobotController.cpp
@@ -1,9 +1,27 @@
 /*
- * Copyright (C) 2019 Istituto Italiano di Tecnologia (IIT)
+ * Copyright (C) 2020 Istituto Italiano di Tecnologia (IIT)
  * All rights reserved.
+ *
+ * This project is dual licensed under LGPL v2.1+ or Apachi License.
+ *
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
  *
  * This software may be modified and distributed under the terms of the
  * GNU Lesser General Public License v2.1 or any later version.
+ *
+ * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
  */
 
 #include "RobotController.h"
@@ -22,7 +40,7 @@
 #include <string>
 
 using namespace gympp::gazebo;
-using namespace gympp::plugins;
+using namespace scenario::plugins::gazebo;
 
 // ====
 // IMPL
@@ -63,7 +81,7 @@ gympp::base::RobotPtr RobotController::Impl::getRobotPtr(const std::string& robo
 
 RobotController::RobotController()
     : System()
-    , pImpl{new Impl(), [](Impl* impl) { delete impl; }}
+    , pImpl{new Impl()}
 {}
 
 RobotController::~RobotController()
@@ -140,7 +158,7 @@ void RobotController::PreUpdate(const ignition::gazebo::UpdateInfo& info,
     }
 }
 
-IGNITION_ADD_PLUGIN(gympp::plugins::RobotController,
-                    gympp::plugins::RobotController::System,
-                    gympp::plugins::RobotController::ISystemConfigure,
-                    gympp::plugins::RobotController::ISystemPreUpdate)
+IGNITION_ADD_PLUGIN(scenario::plugins::gazebo::RobotController,
+                    scenario::plugins::gazebo::RobotController::System,
+                    scenario::plugins::gazebo::RobotController::ISystemConfigure,
+                    scenario::plugins::gazebo::RobotController::ISystemPreUpdate)
