+++ ignition/CMakeLists.txt
@@ -2,29 +2,6 @@
 # This software may be modified and distributed under the terms of the
 # GNU Lesser General Public License v2.1 or any later version.
 
-# =============
-# ECM SINGLETON
-# =============
-
-add_library(ECMSingleton
-    include/gympp/gazebo/ECMSingleton.h
-    src/ECMSingleton.cpp)
-
-target_include_directories(ECMSingleton PUBLIC
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-    $<INSTALL_INTERFACE:${CMAKE_INSTALL_INCLUDEDIR}>)
-
-target_link_libraries(ECMSingleton
-    PUBLIC
-    ignition-gazebo3::core
-    PRIVATE
-    gympp)
-
-if(NOT CMAKE_BUILD_TYPE STREQUAL "PyPI")
-    set_target_properties(ECMSingleton PROPERTIES
-        PUBLIC_HEADER include/gympp/gazebo/ECMSingleton.h)
-endif()
-
 # ==============
 # GAZEBO WRAPPER
 # ==============
@@ -58,7 +35,6 @@ if(NOT CMAKE_BUILD_TYPE STREQUAL "PyPI")
     install(
         TARGETS
         GazeboWrapper
-        ECMSingleton
         EXPORT gympp
         LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
         ARCHIVE DESTINATION ${CMAKE_INSTALL_LIBDIR}
